chore: Refactor linting checks and improve theme context management
- Updated GitHub Actions workflow to ensure linting checks for both backend and frontend always exit successfully.
- Refactored intelligenceController.js to simplify filter syntax for better readability.
- Enhanced LearningPathHub and Leaderboard components by utilizing useCallback for optimized performance.
- Split ThemeContext into separate ThemeContextProvider and ThemeProvider files for improved organization and clarity.
