+++ backend/tests/review.test.js
@@ -1,7 +1,7 @@
 import mongoose from "mongoose";
 import request from "supertest";
 
-jest.mock('redis', () => ({
+jest.mock("redis", () => ({
   createClient: jest.fn(() => ({
     on: jest.fn(),
     connect: jest.fn().mockResolvedValue(null),
